refactor: database log storage logic seqaration
